Modify(sangjun/comment): modal 사용없이 detailcomment 페이지를 만든 버전
