+++ app/src/navigations/MainStack.js
@@ -20,6 +20,7 @@ import {
   SaleList,
   ProfileUpdate,
   Inquiry,
+  DetailComment,
 } from "../screens";
 
 const Stack = createStackNavigator();
@@ -91,6 +92,7 @@ const MainStack = () => {
       <Stack.Screen name="SaleList" component={SaleList} />
       <Stack.Screen name="ProfileUpdate" component={ProfileUpdate} />
       <Stack.Screen name="Inquiry" component={Inquiry} />
+      <Stack.Screen name="DetailComment" component={DetailComment} />
       <Stack.Screen
         name="Login"
         component={Login}
